refactor(test-helper): share reused code
